+++ ChunkyMode/ChunkyRunInfo.cs
@@ -12,35 +12,42 @@ namespace HDeMods {
 		public static ChunkySaveData saveData;
 		
 		// These are to prevent changing settings mid run
-		[SyncVar]
-		public bool doGoldThisRun;
-		[SyncVar]
-		public bool doNerfsThisRun;
-		[SyncVar]
-		public bool doHealBuffThisRun;
-		[SyncVar]
-		public bool doEnemyBoostThisRun;
-		[SyncVar]
-		public int enemyChanceToYapThisRun;
-		[SyncVar]
-		public float enemyYapCooldownThisRun;
-		[FormerlySerializedAs("experimentLimitPestsThisRun")] [SyncVar]
-		public bool limitPestsThisRun;
-		[FormerlySerializedAs("experimentLimitPestsAmountThisRun")] [SyncVar]
-		public float limitPestsAmountThisRun;
+		[FormerlySerializedAs("doNerfsThisRun")] [SyncVar]
+		public bool doEnemyNerfs;
+		
+		[FormerlySerializedAs("doGoldThisRun")] public bool doGoldPenalty;
+		[FormerlySerializedAs("doHealBuffThisRun")] public bool doHealingBuffs;
+		[FormerlySerializedAs("doEnemyBoostThisRun")] public bool doEnemyLimitBoost;
+		[FormerlySerializedAs("enemyChanceToYapThisRun")] public int enemyChanceToYap;
+		[FormerlySerializedAs("enemyYapCooldownThisRun")] public float enemyYapCooldown;
+		[FormerlySerializedAs("limitPestsThisRun")] public bool limitPest;
+		[FormerlySerializedAs("limitPestsAmountThisRun")] public float limitPestAmount;
+		public float rexHealOverride;
+		public float acridHealOverride;
+		public float chirrHealOverride;
+		public float aliemHealOverride;
+		public float submarinerHealOverride;
+		public float ravagerHealOverride;
+		
 
 		public void Awake() {
             instance = this;
             DontDestroyOnLoad(this);
             if (!preSet) return;
-            doEnemyBoostThisRun = saveData.doEnemyBoostThisRun;
-            doHealBuffThisRun = saveData.doHealBuffThisRun;
-            doGoldThisRun = saveData.doGoldThisRun;
-            doNerfsThisRun = saveData.doNerfsThisRun;
-            enemyChanceToYapThisRun = saveData.enemyChanceToYapThisRun;
-            enemyYapCooldownThisRun = saveData.enemyYapCooldownThisRun;
-            limitPestsThisRun = saveData.limitPestsThisRun;
-            limitPestsAmountThisRun = saveData.limitPestsAmountThisRun;
+            doEnemyLimitBoost = saveData.doEnemyLimitBoost;
+            doHealingBuffs = saveData.doHealingBuffs;
+            doGoldPenalty = saveData.doGoldPenalty;
+            doEnemyNerfs = saveData.doEnemyNerfs;
+            enemyChanceToYap = saveData.enemyChanceToYap;
+            enemyYapCooldown = saveData.enemyYapCooldown;
+            limitPest = saveData.limitPest;
+            limitPestAmount = saveData.limitPestAmount;
+            rexHealOverride = saveData.rexHealOverride;
+            acridHealOverride = saveData.acridHealOverride;
+            chirrHealOverride = saveData.chirrHealOverride;
+            aliemHealOverride = saveData.aliemHealOverride;
+            submarinerHealOverride = saveData.submarinerHealOverride;
+            ravagerHealOverride = saveData.ravagerHealOverride;
 		}
 	}
 
@@ -48,21 +55,33 @@ namespace HDeMods {
 		[DataMember(Name = "validCheck")]
 		public bool isValidSave;
 		[DataMember(Name = "gold")]
-		public bool doGoldThisRun;
+		public bool doGoldPenalty;
 		[DataMember(Name = "nerf")]
-		public bool doNerfsThisRun;
+		public bool doEnemyNerfs;
 		[DataMember(Name = "heal")]
-		public bool doHealBuffThisRun;
+		public bool doHealingBuffs;
 		[DataMember(Name = "boost")]
-		public bool doEnemyBoostThisRun;
+		public bool doEnemyLimitBoost;
 		[DataMember(Name = "yapChance")]
-		public int enemyChanceToYapThisRun;
+		public int enemyChanceToYap;
 		[DataMember(Name = "yapTime")]
-		public float enemyYapCooldownThisRun;
+		public float enemyYapCooldown;
 		[DataMember(Name = "exLimitPest")]
-		public bool limitPestsThisRun;
+		public bool limitPest;
 		[DataMember(Name = "exPestCount")]
-		public float limitPestsAmountThisRun;
+		public float limitPestAmount;
+		[DataMember(Name = "rexHeal")]
+		public float rexHealOverride;
+		[DataMember(Name = "acridHeal")]
+		public float acridHealOverride;
+		[DataMember(Name = "chirrHeal")]
+		public float chirrHealOverride;
+		[DataMember(Name = "aliemHeal")]
+		public float aliemHealOverride;
+		[DataMember(Name = "submarinerHeal")]
+		public float submarinerHealOverride;
+		[DataMember(Name = "ravagerHeal")]
+		public float ravagerHealOverride;
 	}
 	
 }